--- manifest.yaml
@@ -1,46 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: carmy-backend-service
-spec:
-  selector:
-    app: carmy-backend
-  ports:
-    - protocol: "TCP"
-      port: 4000
-      targetPort: 4000
-  type: LoadBalancer
-
----
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: carmy-backend
-spec:
-  selector:
-    matchLabels:
-      app: carmy-backend
-  replicas: 1
-  template:
-    metadata:
-      labels:
-        app: carmy-backend
-    spec:
-      containers:
-        - name: carmy-backend
-          #image: registry.digitalocean.com/covid-army-registry/carmy-backend:latest
-          image: docker.io/covidarmy/covid-army-backend:latest
-          env:
-            - name: BEARER_TOKEN
-              valueFrom:
-                secretKeyRef:
-                  name: twitter-bearer-token
-                  key: BEARER_TOKEN
-            - name: MONGO_URI
-              valueFrom:
-                secretKeyRef:
-                  name: mongodb-uri
-                  key: MONGO_URI
-          ports:
-            - containerPort: 4000
-              protocol: "TCP"
