+++ package.json
@@ -5,7 +5,7 @@
   "main": "server.js",
   "scripts": {
     "start": "cross-env NODE_ENV=production node server.js",
-    "dev": "nodemon"
+    "dev": "cross-env NODE_ENV=development nodemon -w src --exec node ./src"
   },
   "repository": {
     "type": "git",
