+++ src/lib/parser.js
@@ -1,7 +1,6 @@
-const allCities = require("./data/newAllCities.json");
-const resourceTypes = require("./data/resources.json");
-
-const categoriesObj = require("./data/categories.json");
+const allCities = require("../../data/newAllCities.json");
+const resourceTypes = require("../../data/resources.json");
+const categoriesObj = require("../../data/categories.json");
 
 const normalize = (text) => {
   return text
